# docs/student-linking-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Edvanz — Student↔Teacher Linking (Request/Approval)",
    "version": "1.0",
...
  },
  "servers": [
    {
      "url": "https://app-edvanz-api-prod.azurewebsites.net",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/teacher/student-links/my-code": {
      "get": {
        "tags": ["Teacher — student links"],
        "summary": "My shareable teacher code",
        "description": "The unique, immutable 8-digit code students type to send a link request. Safe to display prominently — a request still needs this teacher's explicit approval. Requires `Student / ViewList` permission (assistants included).",
        "responses": {
          "200": {
            "description": "The code",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TeacherCodeEnvelope" },
                "examples": {
                  "ok": { "value": { "success": true, "code": "Success", "message": "Done successfully", "data": { "teacherCode": "48291057" } } }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/TeacherNotResolved" }
        }
      }
    },
    "/api/teacher/student-links/requests": {
      "get": {
        "tags": ["Teacher — student links"],
        "summary": "Pending link requests inbox",
        "description": "Newest first. Each item shows what the student typed (`requestedStudentName`, `requestedStudentCode`) plus their real account identity (`studentFullName`, `studentPhoneNumber`, `studentAccountCode`), and — when the typed code matched a roster record — `suggestedMatch` for one-tap accept. `suggestedMatch.isAlreadyLinked = true` warns that accepting with that record will 409. Requires `Student / ViewList`.",
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/pageSize" }
        ],
        "responses": {
          "200": {
            "description": "Paginated pending requests",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LinkRequestPageEnvelope" },
                "examples": {
                  "onePending": {
                    "value": {
                      "success": true,
                      "code": "Success",
                      "message": "Done successfully",
                      "data": {
                        "totalCount": 1,
                        "page": 1,
                        "pageSize": 20,
                        "totalPages": 1,
                        "data": [
                          {
                            "linkId": 12,
                            "requestedStudentName": "Mohamed Ahmed",
                            "requestedStudentCode": "A5",
                            "requestedAt": "2026-07-12T18:40:00Z",
                            "studentAccountCode": "STU4X7B2KM",
                            "studentFullName": "Mohamed Ahmed Ali",
                            "studentPhoneNumber": "01012345678",
                            "suggestedMatch": {
                              "teacherStudentId": 88,
                              "studentName": "Mohamed Ahmed",
                              "studentCode": "A5",
                              "isAlreadyLinked": false
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/TeacherNotResolved" }
        }
      }
    },
    "/api/teacher/student-links/requests/{linkId}/accept": {
      "post": {
        "tags": ["Teacher — student links"],
        "summary": "Accept a request (binds a roster student)",
        "description": "Activates the link and binds it to ONE of this teacher's TeacherStudent roster records — attendance, payments, homework, exams and videos all resolve through that record, so the binding is mandatory. Pass `teacherStudentId` to select explicitly (e.g. from `suggestedMatch` or a roster picker); omit/null to auto-match by the request's typed student code. The student is push-notified. Requires `Student / Add`.",
        "parameters": [{ "$ref": "#/components/parameters/linkId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AcceptLinkRequestRequest" },
              "examples": {
                "autoMatch": { "summary": "Auto-match by the typed student code", "value": { "teacherStudentId": null } },
                "explicit": { "summary": "Explicit roster selection", "value": { "teacherStudentId": 88 } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted (`LinkRequestAccepted`) — returns the new linked-student row",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LinkedStudentEnvelope" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "Request not found (`LinkRequestNotFound`) or selected roster record not found (`RosterStudentNotFound`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          },
          "409": {
            "description": "Already accepted/rejected (`LinkRequestAlreadyResolved`) or roster record already claimed by another account (`RosterStudentAlreadyClaimed`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          },
          "422": {
            "description": "No roster record could be resolved — the teacher must pick one (`RosterStudentSelectionRequired`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          }
        }
      }
    },
    "/api/teacher/student-links/requests/{linkId}/reject": {
      "post": {
        "tags": ["Teacher — student links"],
        "summary": "Reject a request",
        "description": "Marks the request `Rejected` (kept for history — the student sees the outcome and may send a new request later) and push-notifies the student. No body. Requires `Student / Add`.",
        "parameters": [{ "$ref": "#/components/parameters/linkId" }],
        "responses": {
          "200": {
            "description": "Rejected (`LinkRequestRejected`, `data: true`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BooleanEnvelope" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "Request not found (`LinkRequestNotFound`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          },
          "409": {
            "description": "Already accepted/rejected (`LinkRequestAlreadyResolved`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          }
        }
      }
    },
    "/api/teacher/student-links": {
      "get": {
        "tags": ["Teacher — student links"],
        "summary": "My linked students",
        "description": "All `Active` links, newest first — the student's account identity plus the bound roster record (`rosterStudent*` are null if the teacher later deleted that record). Requires `Student / ViewList`.",
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/pageSize" }
        ],
        "responses": {
          "200": {
            "description": "Paginated linked students",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LinkedStudentPageEnvelope" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/TeacherNotResolved" }
        }
      }
    },
    "/api/teacher/student-links/remove": {
      "post": {
        "tags": ["Teacher — student links"],
        "summary": "Remove linked students (one or many)",
        "description": "Soft-removes the given links (state `RemovedByTeacher`); each affected student is push-notified. Ids not owned by this teacher or not currently Active are skipped and echoed back in `skippedLinkIds`. Requires `Student / Delete`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RemoveLinkedStudentsRequest" },
              "examples": {
                "single": { "value": { "linkIds": [12] } },
                "bulk": { "value": { "linkIds": [12, 13, 14] } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Removal outcome (`LinkedStudentsRemoved`)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RemoveResultEnvelope" },
                "examples": {
                  "partial": {
                    "value": {
                      "success": true,
                      "code": "LinkedStudentsRemoved",
                      "message": "Selected students have been removed from your linked students",
                      "data": { "removedCount": 2, "skippedLinkIds": [14] }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Empty selection (`NoLinksSelected`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "None of the ids matched an Active link of this teacher (`LinkNotFound`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "default": 1, "minimum": 1 }
      },
      "pageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
      },
      "linkId": {
        "name": "linkId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" },
        "description": "The link/request row id (`linkId` from the requests inbox)."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "JWT missing or expired",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Forbidden": {
        "description": "Caller lacks the required role/permission",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
...
      "TeacherNotResolved": {
        "description": "Teacher record not found for the authenticated user",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "code": { "type": "string", "description": "Stable, language-independent key to branch on (e.g. LinkRequestAlreadyPending)" },
          "message": { "type": "string", "description": "Localized (Accept-Language: en | ar)" }
        }
      },
      "BooleanEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": { "type": "boolean", "example": true }
        }
      },
...
      "TeacherCodeEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": { "teacherCode": { "type": "string", "example": "48291057" } }
          }
        }
      },
      "RosterStudentSuggestion": {
        "type": "object",
        "nullable": true,
        "properties": {
          "teacherStudentId": { "type": "integer", "format": "int64" },
          "studentName": { "type": "string", "description": "Name on the teacher's roster record" },
          "studentCode": { "type": "string", "example": "A5" },
          "isAlreadyLinked": { "type": "boolean", "description": "True = another account already claims this record; accepting with it will 409" }
        }
      },
      "LinkRequestListItem": {
        "type": "object",
        "properties": {
          "linkId": { "type": "integer", "format": "int64" },
          "requestedStudentName": { "type": "string", "nullable": true },
          "requestedStudentCode": { "type": "string", "nullable": true },
          "requestedAt": { "type": "string", "format": "date-time", "nullable": true },
          "studentAccountCode": { "type": "string" },
          "studentFullName": { "type": "string", "description": "The requesting ACCOUNT's registered name" },
          "studentPhoneNumber": { "type": "string", "nullable": true },
          "suggestedMatch": { "$ref": "#/components/schemas/RosterStudentSuggestion" }
        }
      },
      "LinkRequestPage": {
        "type": "object",
        "properties": {
          "totalCount": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/LinkRequestListItem" } }
        }
      },
      "LinkRequestPageEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/LinkRequestPage" }
        }
      },
      "AcceptLinkRequestRequest": {
        "type": "object",
        "properties": {
          "teacherStudentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Explicit roster record to bind. Omit/null to auto-match by the request's typed student code. If neither resolves → 422."
          }
        }
      },
      "LinkedStudentListItem": {
        "type": "object",
        "properties": {
          "linkId": { "type": "integer", "format": "int64" },
          "linkedAt": { "type": "string", "format": "date-time" },
          "studentAccountCode": { "type": "string" },
          "studentFullName": { "type": "string" },
          "studentPhoneNumber": { "type": "string", "nullable": true },
          "teacherStudentId": { "type": "integer", "format": "int64", "nullable": true, "description": "Null if the teacher deleted the roster record after linking" },
          "rosterStudentName": { "type": "string", "nullable": true },
          "rosterStudentCode": { "type": "string", "nullable": true }
        }
      },
      "LinkedStudentEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string", "example": "LinkRequestAccepted" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/LinkedStudentListItem" }
        }
      },
      "LinkedStudentPage": {
        "type": "object",
        "properties": {
          "totalCount": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/LinkedStudentListItem" } }
        }
      },
      "LinkedStudentPageEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/LinkedStudentPage" }
        }
      },
      "RemoveLinkedStudentsRequest": {
        "type": "object",
        "required": ["linkIds"],
        "properties": {
          "linkIds": {
            "type": "array",
            "minItems": 1,
            "items": { "type": "integer", "format": "int64" },
            "description": "One or many link ids from the linked-students list"
          }
        }
      },
      "RemoveResultEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "code": { "type": "string", "example": "LinkedStudentsRemoved" },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "removedCount": { "type": "integer" },
              "skippedLinkIds": { "type": "array", "items": { "type": "integer", "format": "int64" } }
            }
          }
        }
      }
    }
  }
}